@@ -1,13 +1,13 @@
 package com.gy.motionlayout
 
-import android.support.constraint.ConstraintLayout
-import android.support.v7.widget.RecyclerView
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 
-class DemoAdapter(private val demoList: ArrayList<DemoAdapter.Demo>) : RecyclerView.Adapter<DemoAdapter.ViewHolder>() {
+class DemoAdapter(private val demoList: ArrayList<DemoAdapter.Demo>) : androidx.recyclerview.widget.RecyclerView.Adapter<DemoAdapter.ViewHolder>() {
 
 
     override fun getItemCount(): Int {
@@ -27,7 +27,7 @@ class DemoAdapter(private val demoList: ArrayList<DemoAdapter.Demo>) : RecyclerV
     }
 
 
-    class ViewHolder(val layout :ConstraintLayout) : RecyclerView.ViewHolder(layout) {
+    class ViewHolder(val layout : ConstraintLayout) : androidx.recyclerview.widget.RecyclerView.ViewHolder(layout) {
         val item = layout.findViewById<TextView>(R.id.item)
         var layoutId = 0
 
