@@ -2,17 +2,19 @@ package com.gy.motionlayout.example
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
 import android.widget.Toast
+import com.gy.motionlayout.CarouselActivity
 import com.gy.motionlayout.R
+import com.gy.motionlayout.example.activity.*
 
 class DemoActivity : AppCompatActivity(){
 
     private lateinit var mAdapter : MainAdapter
-    private lateinit var mLayoutManager : LinearLayoutManager
+    private lateinit var mLayoutManager : androidx.recyclerview.widget.LinearLayoutManager
 
     private val mAdapterData : Array<MainAdapter.Demo> = arrayOf(
             MainAdapter.Demo("Basic Collapsing Toolbar", ExampleTypes.DEFAULT, R.layout.collapsing_toolbar),
@@ -20,21 +22,23 @@ class DemoActivity : AppCompatActivity(){
             MainAdapter.Demo("Collapsing Toolbar w/ Cover Example", ExampleTypes.FULLSCREEN, R.layout.collapsing_toolbar_with_cover),
             MainAdapter.Demo("Complex Animation Example", ExampleTypes.WITHOUT_RECYCLER_VIEW, R.layout.complex_animation_example, ComplexAnimationActivity::class.java),
             MainAdapter.Demo("Multiple Animation Example", ExampleTypes.FULLSCREEN, R.layout.multiple_animation_example, ComplexAnimationActivity::class.java),
-            MainAdapter.Demo("Basic Keyframe Example", ExampleTypes.WITHOUT_RECYCLER_VIEW, R.layout.basic_key_frame_example)
-//            MainAdapter.Demo("Basic Keyframe Example 2", ExampleTypes.WITHOUT_RECYCLER_VIEW, R.layout.basic_key_frame_example_2),
-//            MainAdapter.Demo("Notification Center Example", ExampleTypes.DEFAULT, R.layout.notification_center_example),
-//            MainAdapter.Demo("ViewPager Example", ExampleTypes.VIEW_PAGER, R.layout.viewpager_example, ViewPagerActivity::class.java),
-//            MainAdapter.Demo("Youtube Fragment Transition Example", ExampleTypes.FRAGMENT, R.layout.youtube_fragment_transition_example, FragmentActivity::class.java)
+            MainAdapter.Demo("CarouselDemo1", ExampleTypes.FULLSCREEN, R.layout.demo1, CarouselActivity::class.java),
+            MainAdapter.Demo("AnimationActivity", ExampleTypes.FULLSCREEN, 0, AnimationActivity::class.java),
+            MainAdapter.Demo("AnimationCardActivity", ExampleTypes.FULLSCREEN, 0, AnimationCardActivity::class.java),
+            MainAdapter.Demo("AnimationCarActivity", ExampleTypes.FULLSCREEN, 0, AnimationCarActivity::class.java),
+            MainAdapter.Demo("模仿华为拨号键", ExampleTypes.FULLSCREEN, 0, HuaweiTelActivity::class.java),
+            MainAdapter.Demo("android 11 彩蛋", ExampleTypes.FULLSCREEN, 0, EasterEggs11Activity::class.java),
+            MainAdapter.Demo("跟其他组件配合使用", ExampleTypes.FULLSCREEN, 0, ViewPagerDemoActivity::class.java)
     )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_demo)
 
-        val recyclerView = findViewById<RecyclerView>(R.id.recyclerView)
+        val recyclerView = findViewById<androidx.recyclerview.widget.RecyclerView>(R.id.recyclerView)
 
         mAdapter = MainAdapter(mAdapterData)
-        mLayoutManager = LinearLayoutManager(this)
+        mLayoutManager = androidx.recyclerview.widget.LinearLayoutManager(this)
         recyclerView.apply {
             adapter = mAdapter
             layoutManager = mLayoutManager
@@ -44,16 +48,12 @@ class DemoActivity : AppCompatActivity(){
 
     fun start(activity:Class<*>, layoutFileId: Int, types: ExampleTypes?, position: Int){
 
-        if (position > 4) {
-            Toast.makeText(this, "Coming soon...", Toast.LENGTH_LONG).show()
-        } else {
-            Log.i("cccccccc","class=="+activity.simpleName)
-            val intent = Intent(this, activity).apply {
-                putExtra("layoutId", layoutFileId)
-                putExtra("type", types?.ordinal ?: ExampleTypes.DEFAULT.ordinal)
-            }
-            startActivity(intent)
+        Log.i("cccccccc","class=="+activity.simpleName)
+        val intent = Intent(this, activity).apply {
+            putExtra("layoutId", layoutFileId)
+            putExtra("type", types?.ordinal ?: ExampleTypes.DEFAULT.ordinal)
         }
+        startActivity(intent)
     }
 
 }
